+++ src/components/layouts/navigation.tsx
@@ -1,49 +1,13 @@
 "use client";
 
 import React, { useState } from "react";
-import { Sidebar, SidebarBody, SidebarLink } from "@/components/ui/sidebar";
-import { IconArrowLeft, IconBrandTabler, IconSettings, IconUserBolt } from "@tabler/icons-react";
-import Link from "next/link";
-import Image from "next/image";
+import { Sidebar, SidebarBody } from "@/components/ui/sidebar";
+import { Logo } from "@/components/navigation/Logo";
+import { ProfileLink } from "@/components/navigation/ProfileLink";
+import { SidebarLinks } from "@/components/navigation/SidebarLinks";
 import { cn } from "@/lib/utils";
-import profile from "../../../public/images/profile.jpeg";
 
-export default function Navigation({
-    children,
-}: Readonly<{
-    children: React.ReactNode;
-}>) {
-    const links = [
-        {
-            label: "Dashboard",
-            href: "/menu/dashboard",
-            icon: (
-                <IconBrandTabler className="text-neutral-700 dark:text-neutral-200 h-5 w-5 flex-shrink-0" />
-            ),
-        },
-        {
-            label: "Profile",
-            href: "/menu/profile",
-            icon: (
-                <IconUserBolt className="text-neutral-700 dark:text-neutral-200 h-5 w-5 flex-shrink-0" />
-            ),
-        },
-        {
-            label: "Settings",
-            href: "/settings",
-            icon: (
-                <IconSettings className="text-neutral-700 dark:text-neutral-200 h-5 w-5 flex-shrink-0" />
-            ),
-        },
-        {
-            label: "Logout",
-            href: "/logout",
-            icon: (
-                <IconArrowLeft className="text-neutral-700 dark:text-neutral-200 h-5 w-5 flex-shrink-0" />
-            ),
-        },
-    ];
-
+export default function Navigation({ children }: { children: React.ReactNode }) {
     const [open, setOpen] = useState(true);
 
     return (
@@ -52,29 +16,9 @@ export default function Navigation({
                 <SidebarBody className="justify-between gap-10">
                     <div className="flex flex-col flex-1 overflow-y-auto overflow-x-hidden">
                         <Logo />
-                        <div className="mt-8 flex flex-col gap-2">
-                            {links.map((link, idx) => (
-                                <SidebarLink key={idx} link={link} />
-                            ))}
-                        </div>
+                        <SidebarLinks />
                     </div>
-                    <div>
-                        <SidebarLink
-                            link={{
-                                label: "Manu Arora",
-                                href: "#",
-                                icon: (
-                                    <Image
-                                        src={profile}
-                                        className="h-7 w-7 flex-shrink-0 rounded-full"
-                                        width={50}
-                                        height={50}
-                                        alt="Avatar"
-                                    />
-                                ),
-                            }}
-                        />
-                    </div>
+                    <ProfileLink />
                 </SidebarBody>
             </Sidebar>
             <main className="flex-1 overflow-auto dark:bg-neutral-900">
@@ -83,18 +27,3 @@ export default function Navigation({
         </div>
     );
 }
-
-export const Logo = () => {
-    return (
-        <Link
-            href="/"
-            className="font-normal flex space-x-2 items-center text-sm text-black py-1 relative z-20"
-        >
-            <div className="h-5 w-6 bg-black dark:bg-white rounded-br-lg rounded-tr-sm rounded-tl-lg rounded-bl-sm flex-shrink-0" />
-            <span className="font-medium text-black dark:text-white whitespace-pre">
-                Acet Labs
-            </span>
-        </Link>
-    );
-};
-
